Separate trigger / orchestrator
